packages/python/jango/requirements.txt
Django==3.1.7
djangorestframework==3.12.0
environ==1.0
drf-yasg==1.20.0